requirements/local.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
Pillow==8.4.0
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
Unipath==1.1
yapf==0.31.0